Add new functions to Purchase UI: Update, load, and add new purchase

+++ src/main/java/PurchaseModel.java
@@ -2,7 +2,7 @@ public class PurchaseModel {
     public int mPurchaseId;
     public int mProductID;
     public int mCustomerID;
-    public double mQuantity;
+    public int mQuantity;
 
     public String toString() {
         StringBuilder purchaseSB = new StringBuilder("(");
